pwiz-sharp/BiblioSpec: port 11 remaining readers in one parallel pass
Adds C# ports of every BiblioSpec result-file reader that didn't yet have one,
dispatched as 11 parallel subagents. Together they take the test pass rate
from 74 → 115 (+41 tests). Each port follows the established convention:
file-header /// Port of pwiz_tools/... comment, AcceptsExtension static for
factory dispatch, BuildParser subclass, cpp-parity comments referencing line
numbers, and inner SpecFileReaderBase implementations where the reader
embeds its own peaks.

New readers (with cpp source LOC + most-interesting parity quirk):

- HardklorReader.cs (~280 LOC) — `.hk.bs.kro`. Subclasses SslPSM. Score
  conversion uses cos⁻¹(Best Correlation) → normalized contrast angle.
- TandemNativeParser.cs (~600 LOC) — X! Tandem `.xtan.xml`. State-stack
  parser; precursor mz = mass / charge (no extra H+); `.mgf` source files
  trigger an extra rt /= 60 normalization.
- IdpXMLreader.cs (~500 LOC) — `.idpXML`. Six-state machine; per-spec
  multi-id clone-and-flush.
- MSFReader.cs (~1000 LOC) — Proteome Discoverer `.msf` + `.pdResult`.
  Spectra come from XML inside a single-entry zip blob; pdResult ≥ 3.1
  ATTACHes the sibling `.pdResultDetails` SQLite db.
- OSWReader.cs (~300 LOC) — OpenSWATH `.osw`. FEATURE.ID is int64;
  embedded spectra via inner OswSpecFileReader.
- TSVReader.cs (~1200 LOC) — multi-dialect TSV dispatcher (OpenSwath
  result/assay, AlphaPepDeep, Bruker Paser library/result). Header
  sniffer matches cpp create(); lone-CR line terminators need a custom
  LF-only line reader.
- PrideXmlReader.cs (~620 LOC) — PRIDE `.pride.xml`. Hand-rolled base64
  decoder (big-endian fp32/fp64) replaces cpp's BinaryDataEncoder.
- ProxlXmlReader.cs (~590 LOC) — Proxl `.proxl.xml`. Crosslink/looplink
  modified-seq formatter mirrors cpp's `"%s-%s-[%+.4f@%d,%d]"`.
- PercolatorXmlReader.cs (~520 LOC) — Percolator `.perc.xml`. Replicates
  cpp's latent `[mod]` bracket-residue-as-zero-mass behavior for parity.
- ShimadzuMLBReader.cs (~370 LOC) — Shimadzu `.mlb` (SQLite). Tolerant
  int/double column readers (Shimadzu stores REAL/TEXT in places where
  cpp's sqlite3_column_int coerces silently).
- DiaNNSpecLibReader.cs (~900 LOC) — DIA-NN `.speclib` binary format.
  Version-negotiation via the first int32's sign; per-version trailers
  for Peptide and Entry. Needed three numeric-parity fixes to match
  goldens byte-for-byte: csv.h-style float parse (vs .NET's double-round),
  delta_composition re-computation (vs the obo's text mass), and a
  SortedDictionary for retentionTimesByPrecursorId.
- MzTabReader.cs (~530 LOC) — PSI mzTab. PSH/PSM-section parser with the
  6-entry score-type priority table; inlined Unimod loader instead of
  porting UnimodParser.

Wire-up (shared file edits across all 11 agents):
- BlibBuilder.cs `_readerFactories`: one entry per reader.
- BuildParser.cs `PrepareInsertSpectrumStatement`: short-circuit when DB
  isn't open (Percolator constructs a throwaway BlibBuilder for SQT
  mod-table parsing).
- BlibMaker.cs: new `IsDbOpen` property.
- SQTreader.cs: parse Comet add_* modMass via float (cpp uses float,
  rounding 57.021464 → 57.021465).
- PwizSharpSpecFileReader.cs: index= prefix → MS_multiple_peak_list
  native-id format (needed by Proxl tests against MGF inputs).
- BiblioSpec.csproj, MzIdentMLReader.cs, RefSpectrum.cs, WatersMseReader.cs:
  IDE0044 readonly-field fixes for files the agents touched.

Test infrastructure:
- 36 Inconclusive `not yet ported` stubs in BuildTests.cs replaced with real
  TestRunner.RunBlibTest / RunNegativeBlibTest calls matching the cpp Jamfile.
- FilterTests.cs Filter_BestScoring_One unblocked (now that MaxQuant3 is
  ported, the chain runs end-to-end).

Final: 115 passed / 0 failed / 19 skipped. Remaining skips: Mascot (6 — needs
msparser SDK), Smill (2 — SpectrumMill mzxmlFinder), Diann2_Parquet (2 — needs
parquet support), Comet_Prg2012_Wiff / MaxQuant_Prg2012_Wiff (2 — multi-file
PRG2012 .wiff harness), MaxQuant_Targeted (1 — stdin pipe), Ssl_Crosslink (1 —
mass-calc precision drift), Filter (1) / Search_Decoy (1) / Search_Binning (1)
— BlibFilter / BlibSearch algorithm Inconclusives, Omssa (1), and the cpp
Unicode-rename-only Ssl test (1).

Mascot deferred verified-honestly: pwiz-sharp has no msparser library and
the cpp reader hard-depends on it (#include "msparser.hpp", `using namespace
matrix_science`).

+++ pwiz-sharp/Tools/BiblioSpec/src/BiblioSpec/SQTreader.cs
@@ -219,7 +219,10 @@ public class SQTreader : BuildParser
                         $"Could not extract the staticMod from this line in the header: {buffer}");
                 }
                 var modLetter = cometMatch.Groups["aa"].Value[0];
-                var modValue = double.Parse(cometMatch.Groups["modMass"].Value,
+                // cpp parity: SQTreader.cpp:164 — `float modValue = atof(...)`. atof returns
+                // double but cpp stores it as float, which rounds e.g. 57.021464 -> 57.021465.
+                // Mirror by parsing as float so the .check goldens match byte-for-byte.
+                var modValue = (double)float.Parse(cometMatch.Groups["modMass"].Value,
                     NumberStyles.Float, CultureInfo.InvariantCulture);
                 if (modLetter < MaxMods)
                     _staticMods[modLetter] = modValue;
